pyproject.toml
[project]
name = "tcc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "deap>=1.4.2",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.0",
    "pandas>=2.2.3",
    "pyyaml>=6.0.2",
    "ruff>=0.9.2",
    "seaborn>=0.13.2",
]
